+++ src/main/java/net/frozenblock/wilderwild/tag/WilderBiomeTags.java
@@ -169,7 +169,6 @@ public final class WilderBiomeTags {
 	public static final TagKey<Biome> HAS_COARSE_DIRT_PILE_WITH_DISK = bind("has_coarse_dirt_pile_with_disk");
 	public static final TagKey<Biome> HASS_COBBLESTONE_PILE = bind("has_cobblestone_pile");
 	public static final TagKey<Biome> HAS_COARSE_DIRT_TRANSITION_DISK = bind("has_coarse_dirt_transition_disk");
-	public static final TagKey<Biome> HAS_PODZOL_TRANSITION_DISK = bind("has_podzol_transition_disk");
 	public static final TagKey<Biome> HAS_DECORATIVE_MUD = bind("has_decorative_mud");
 	public static final TagKey<Biome> HAS_PACKED_MUD_ORE = bind("has_packed_mud_ore");
 	public static final TagKey<Biome> HAS_COARSE_DIRT_PATH = bind("has_coarse_dirt_path");