+++ src/main/java/net/frozenblock/wilderwild/tag/WilderBlockTags.java
@@ -127,7 +127,6 @@ public final class WilderBlockTags {
 	public static final TagKey<Block> NETHER_GEYSER_REPLACEABLE = bind("nether_geyser_replaceable");
 	public static final TagKey<Block> OASIS_PATH_REPLACEABLE = bind("oasis_path_replaceable");
 	public static final TagKey<Block> COARSE_DIRT_DISK_REPLACEABLE = bind("coarse_dirt_disk_replaceable");
-	public static final TagKey<Block> PODZOL_DISK_REPLACEABLE = bind("podzol_disk_replaceable");
 	public static final TagKey<Block> MESOGLEA = bind("mesoglea");
 	public static final TagKey<Block> NEMATOCYSTS = bind("nematocysts");