+++ src/main/java/net/frozenblock/wilderwild/world/feature/configured/WilderMiscConfigured.java
@@ -147,7 +147,6 @@ public final class WilderMiscConfigured {
 	// DYING FOREST
 	public static final FrozenConfiguredFeature<FadingDiskTagFeatureConfig, ConfiguredFeature<FadingDiskTagFeatureConfig, ?>> COARSE_DIRT_DISK_AND_PILE = register("coarse_dirt_disk_and_pile");
 	public static final FrozenConfiguredFeature<FadingDiskTagFeatureConfig, ConfiguredFeature<FadingDiskTagFeatureConfig, ?>> COARSE_TRANSITION_DISK = register("coarse_dirt_transition_disk");
-	public static final FrozenConfiguredFeature<FadingDiskTagFeatureConfig, ConfiguredFeature<FadingDiskTagFeatureConfig, ?>> PODZOL_TRANSITION_DISK = register("podzol_transition_disk");
 
 	// SNOW
 	public static final FrozenConfiguredFeature<NoneFeatureConfiguration, ConfiguredFeature<NoneFeatureConfiguration, ?>> SNOW_BLANKET = register("snow_blanket");
@@ -947,22 +946,6 @@ public final class WilderMiscConfigured {
 				true,
 				BlockStateProvider.simple(Blocks.COARSE_DIRT.defaultBlockState()),
 				BlockStateProvider.simple(Blocks.COARSE_DIRT.defaultBlockState()),
-				UniformInt.of(2, 4),
-				0.95F,
-				0.925F,
-				0.65F,
-				0.5F,
-				WilderBlockTags.COARSE_DIRT_DISK_REPLACEABLE,
-				WilderBlockTags.COARSE_DIRT_DISK_REPLACEABLE,
-				Heightmap.Types.OCEAN_FLOOR_WG
-			)
-		);
-
-		PODZOL_TRANSITION_DISK.makeAndSetHolder(FrozenFeatures.FADING_DISK_TAG_FEATURE,
-			new FadingDiskTagFeatureConfig(
-				true,
-				BlockStateProvider.simple(Blocks.PODZOL.defaultBlockState()),
-				BlockStateProvider.simple(Blocks.PODZOL.defaultBlockState()),
 				UniformInt.of(2, 4),
 				0.95F,
 				0.925F,